breaking up scans into separate modules
